Order process fixed, save to database, empty item list solved

+++ DAL/IOrderDB.cs
@@ -7,7 +7,7 @@ namespace DAL
     {
         Order CreateOrder(Order order);
         List<Order> GetOrders();
-        Order GetOrderById(int id);
+        Order GetOrderById(long id);
         int AddOrder(Order order);
         int UpdateOrder(Order order);
         int DeleteOrder(int id);
